solved erased first input

+++ minishell/src/parsing/parsing.c
@@ -6,7 +6,7 @@
 /*   By: bbresil <bbresil@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/05 12:13:29 by bbresil           #+#    #+#             */
-/*   Updated: 2024/01/10 12:15:02 by bbresil          ###   ########.fr       */
+/*   Updated: 2024/01/10 12:28:27 by bbresil          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,8 +61,11 @@ void	keep_last_input(t_cmd **cmd_tab)
 			lex = cmd_tab[i]->input;
 			while (lex && lex->next)
 			{
-				if (unlink(lex->word) != 0)
-					perror("error deleting file");
+				if (lex->type == LIMITER)
+				{
+					if (unlink(lex->word) != 0)
+						perror("error deleting file");
+				}
 				lex = ft_remove_lex_node(&lex, lex);
 			}
 			cmd_tab[i]->input = lex;
@@ -71,6 +74,31 @@ void	keep_last_input(t_cmd **cmd_tab)
 	}
 }
 
+// working
+// void	keep_last_input(t_cmd **cmd_tab)
+// {
+// 	int		i;
+// 	t_lexer	*lex;
+
+// 	i = 0;
+// 	while (cmd_tab[i])
+// 	{
+// 		if (cmd_tab[i]->input)
+// 		{
+// 			lex = cmd_tab[i]->input;
+// 			while (lex && lex->next)
+// 			{
+// 				if (unlink(lex->word) != 0)
+// 					perror("error deleting file");
+// 				lex = ft_remove_lex_node(&lex, lex);
+// 			}
+// 			cmd_tab[i]->input = lex;
+// 		}
+// 		i++;
+// 	}
+// }
+
+//not working
 // void	keep_last_input(t_cmd **cmd_tab)
 // {
 // 	int		i;
